adds support for song duration time
