+++ splitutil.py
@@ -1,4 +1,3 @@
-
 def updateTimeChange(time_elapsed, track_time):
     elapsed_h = int(time_elapsed[:1])
     elapsed_m = int(time_elapsed[2:4])
@@ -34,6 +33,16 @@ def updateTimeChange(time_elapsed, track_time):
         str_s = str(elapsed_s)
 
     return '{}:{}:{}'.format(str(elapsed_h), str_m, str_s)
+
+
+def timeToSeconds(time):
+    parts = time.split(":")
+    seconds = None
+    if len(parts) == 3:  # h:m:s
+        seconds = int(parts[0])*3600 + int(parts[1])*60 + int(parts[2])
+    elif len(parts) == 2:  # m:s
+        seconds = int(parts[0])*60 + int(parts[1])
+    return seconds
 
 
 def writeTracksToFile(track_times, track_titles):
