+++ tracks.txt
@@ -1,11 +1,10 @@
-0:00:00 - Is This It
-0:02:35 - The Modern Age
-0:06:07 - Soma
-0:08:45 - Barely Legal
-0:12:39 - Someday
-0:15:46 - Alone, Together
-0:18:58 - Last Nite
-0:22:16 - Hard to Explain
-0:26:04 - New York City Cops
-0:29:40 - Trying Your Luck
-0:33:08 - Take It or Leave It
+2:35 - Is This It
+3:32 - The Modern Age
+2:38 - Soma
+3:54 - Barely Legal
+3:07 - Someday
+3:12 - Alone, Together
+3:18 - Hard to Explain
+3:36 - New York City Cops
+3:28 - Trying Your Luck
+3:16 - Take it or Leave it
